files (1)/requirements.txt
flask==3.1.1
flask-cors==6.0.2
PyJWT==2.10.1
bcrypt==5.0.0
